requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asyncio==4.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.2
multidict==6.6.4
openai==1.109.0
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-telegram-bot==22.4
redis==6.4.0
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.41
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1

# Additional dependencies for advanced features
python-dotenv==1.0.0
aiofiles==23.2.1
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2
hiredis==2.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pydantic==2.5.2
colorama==0.4.6
pydub==0.25.1
speech-recognition==3.10.0
beautifulsoup4==4.12.2
selenium==4.16.0
asyncio-mqtt==0.16.1
uvloop==0.19.0
structlog==23.2.0
prometheus-client==0.19.0
cryptography==41.0.8
bcrypt==4.1.2
pyyaml==6.0.1
toml==0.10.2
orjson==3.9.10
pytz==2023.3
croniter==2.0.1
pathlib2==2.3.7
watchdog==3.0.0
psutil==5.9.6
python-decouple==3.8
